+++ domestic/models.py
@@ -8,7 +8,13 @@ from django.db import models
 from django.http import Http404
 from great_components.mixins import GA360Mixin
 from modelcluster.fields import ParentalManyToManyField
-from wagtail.admin.edit_handlers import FieldPanel, StreamFieldPanel
+from wagtail.admin.edit_handlers import (
+    FieldPanel,
+    ObjectList,
+    StreamFieldPanel,
+    TabbedInterface,
+    cached_classmethod,
+)
 from wagtail.contrib.table_block.blocks import TableBlock
 from wagtail.core.blocks.field_block import RichTextBlock
 from wagtail.core.blocks.stream_block import StreamBlock, StreamBlockValidationError
@@ -66,11 +72,22 @@ class BaseContentPage(
 ):
     """Minimal abstract base class for pages ported from the V1 Great.gov.uk site"""
 
-    promote_panels = []  #  Hide the Promote panel
+    promote_panels = []
     folder_page = False  # Some page classes will have this set to true to exclude them from breadcrumbs
 
     class Meta:
         abstract = True
+
+    @cached_classmethod
+    def get_edit_handler(cls):  # noqa
+        panels = [
+            # Normal Wagtail panels.
+            ObjectList(cls.content_panels, heading='Content'),
+            # Added custom SEO panels in new tab.
+            ObjectList(SeoMixin.seo_meta_panels, heading='SEO', classname='seo'),
+            ObjectList(cls.settings_panels, heading='Settings', classname='settings'),
+        ]
+        return TabbedInterface(panels).bind_to(model=cls)
 
     def get_ancestors_in_app(self):
         """